WIP add settings to create and edit form
