+++ frontend/src/components/data-products/data-product-form/data-product-form.component.tsx
@@ -1,4 +1,4 @@
-import { Button, Form, FormProps, Input, Popconfirm, Select, Space } from 'antd';
+import { Button, Flex, Form, FormProps, Input, Popconfirm, Select, Space } from 'antd';
 import { useTranslation } from 'react-i18next';
 import styles from './data-product-form.module.scss';
 import {
@@ -27,6 +27,7 @@ import { FORM_GRID_WRAPPER_COLS, MAX_DESCRIPTION_INPUT_LENGTH } from '@/constant
 import { selectFilterOptionByLabelAndValue } from '@/utils/form.helper.ts';
 import { useGetAllDataProductTypesQuery } from '@/store/features/data-product-types/data-product-types-api-slice.ts';
 import { DataProductMembershipRole, DataProductUserMembershipCreateContract } from '@/types/data-product-membership';
+import { DataProductSettings } from '@/components/data-products/data-product-settings/data-product-settings.component.tsx';
 
 type Props = {
     mode: 'create' | 'edit';
@@ -178,7 +179,6 @@ export function DataProductForm({ mode, dataProductId }: Props) {
             });
         }
     }, [currentDataProduct, mode]);
-
     return (
         <Form
             form={form}
@@ -295,6 +295,9 @@ export function DataProductForm({ mode, dataProductId }: Props) {
             >
                 <TextArea rows={4} count={{ show: true, max: MAX_DESCRIPTION_INPUT_LENGTH }} />
             </Form.Item>
+            {/* If mode is create -> no autosave, but save only after the product id is known? */}
+            {/* Editing seemed to work fine, even with autosave? */}
+            <DataProductSettings dataProductId={dataProductId}/>
             <Form.Item>
                 <Space>
                     <Button
